+++ spec/whois/record/nameserver_spec.rb
@@ -1,4 +1,6 @@
 require 'spec_helper'
+require 'whois/record/nameserver'
+
 
 describe Whois::Record::Nameserver do