+++ utils/bm_regexp.rb
@@ -3,7 +3,7 @@ require 'benchmark'
 
 
 Benchmark.bmbm do |x|
-  string = File.read(File.expand_path("../../spec/fixtures/responses/whois.nic.tr/registered.txt", __FILE__))
+  string = File.read(File.expand_path("../../spec/fixtures/responses/whois.nic.tr/status_registered.txt", __FILE__))
   result = "2010-Aug-22."
   x.report("w/i delimiter") do
     100_000.times do